Warn if source announcement filename does not contain "announce"
User had issues with uploading worship file as an announcement file.
Program is not filtering filenames since from time to time a general
slideshow may be used in place of an announcement file (at funerals or
other special events).  New solution is to prompt the user to confirm
the chosen file if the source filename does not contain "announce" as a
substring.